Support for RetainResources
This change was a bit more involved as in order to delete a
Lambda@Edge resource StackDestroy must be called twice. Once
without RetainResources, then again with RetainResources
since the stack _must_ be in a failed state for RetainResources
to work.

https://docs.aws.amazon.com/AWSCloudFormation/latest/APIReference/API_DeleteStack.html

The following functionality were added to support retain_resources;

* The destroy command nows checks for a failed state.
* If the state is failed and `retain_resources` is defined retry the
  destory passing RetainResources otherwise run as before.